requirements.txt
aioredis==1.3.1
asgiref==3.2.10
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.7.1
Automat==20.2.0
cffi==1.14.0
channels==2.4.0
channels-redis==3.0.1
constantly==15.1.0
cryptography==3.0
daphne==2.5.0
dj-database-url==0.5.0
Django==3.0.8
django-cors-headers==3.4.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
hiredis==1.1.0
hyperlink==19.0.0
idna==2.10
incremental==17.5.0
msgpack==1.0.0
Pillow==7.2.0
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-dotenv==0.14.0
pytz==2020.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
whitenoise==5.1.0
zope.interface==5.1.0
